:art: move to new file and use singleton
